archifska/pyproject.toml
[project]
name = "archifska"
version = "0.1.0"
description = "media management for a system with two mountpoints, works with qbittorrent+radarr/sonarr"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "parse-torrent-title>=2.8.1",
    "pyarr>=5.2.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "qbittorrent-api>=2025.7.0",
    "rapidfuzz>=3.14.1",
]
